src/5. Tools and Agents/3_multi_agent_systems/3_Google_A2A_protocol/standalone agents/A2A_standalone_agent_tavily/agent_tests.py
from dotenv import load_dotenv

from agent import Agent

load_dotenv()

agent = Agent()
sessionID = "session_12345"

query = "Tell me abaout Adam Mickiewicz"
response = agent.invoke(query, sessionID)
print(response)
